+++ atcodertools/constprediction/models/problem_constant_set.py
@@ -1,3 +1,4 @@
+from atcodertools.common.judgetype import Judge
 
 
 class ProblemConstantSet:
@@ -6,7 +7,9 @@ class ProblemConstantSet:
                  mod: int = None,
                  yes_str: str = None,
                  no_str: str = None,
+                 judge_method: Judge = None,
                  ):
         self.mod = mod
         self.yes_str = yes_str
         self.no_str = no_str
+        self.judge_method = judge_method