+++ webapp/src/models/Language.ts
@@ -1,6 +1,6 @@
 
-type Language = 'cpp' | 'rust' | 'java' | 'python';
-export const ALL_LANGUAGES : Language[] = ['cpp', 'rust', 'java', 'python'];
+type Language = 'cpp' | 'rust' | 'java' | 'python' | 'd' | 'nim' | 'cs';
+export const ALL_LANGUAGES : Language[] = ['cpp', 'rust', 'java', 'python', 'd', 'nim', 'cs'];
 
 export const langToDisplayName = (language: Language) => {
     switch (language){
@@ -8,6 +8,9 @@ export const langToDisplayName = (language: Language) => {
         case 'java': return 'Java';
         case 'rust': return 'Rust';
         case 'python': return 'Python 3';
+        case 'd': return 'D';
+        case 'nim': return 'Nim';
+        case 'cs': return 'C#';
     }
     throw Error(`no display name for ${language}`);
 };